display temperature decimal point value

+++ Drivers/generalIO/tempsensors.h
@@ -15,12 +15,12 @@
 uint16_t readColdJunctionSensorTemp_mC(void);
 uint16_t coldJunctionTemp_mC_To_uV(int tempX10);
 uint16_t readTipSensorADC_Avg(void);
-uint16_t readTipTemperatureCompensated(uint8_t new);
+float readTipTemperatureCompensated(uint8_t new);
 uint16_t realTempToADC(uint16_t real);
 void setCurrentTip(uint8_t tip);
 tipData * getCurrentTip();
 
-long map(long x, long in_min, long in_max, long out_min, long out_max);
-uint16_t adc2Human(uint16_t);
+float map(float x, float in_min, float in_max, float out_min, float out_max);
+float adc2Human(uint16_t);
 uint16_t human2adc(uint16_t);
 #endif /* GENERALIO_TEMPSENSORS_H_ */
